Refactor client tests.
